Create NoteView
