+++ index.js
@@ -5,5 +5,6 @@
 import { AppRegistry } from 'react-native';
 import App from './App';
 import { name as appName } from './app.json';
-AppRegistry.registerComponent(appName, () => App);
+import TestComponent from './src/components/TestComponent';
+AppRegistry.registerComponent(appName, () => TestComponent);
